refactor(comments): load comments-legacy as the default plugin
