+++ packages/sanity/src/_singletons/context/CommentsEnabledContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'sanity/_createContext'
 
-import type {CommentsEnabledContextValue} from '../../core/comments/context/enabled/types'
+import type {CommentsEnabledContextValue} from '../../core/comments-legacy/context/enabled/types'
 
 // NOTE: We initialize this context with a default value (`enabled: false`)
 // rather than `null` to handle cases where the comments feature's availability
