+++ packages/sanity/src/_singletons/context/CommentsIntentContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'sanity/_createContext'
 
-import type {CommentsIntentContextValue} from '../../core/comments/context/intent/types'
+import type {CommentsIntentContextValue} from '../../core/comments-legacy/context/intent/types'
 
 /**
  * @internal
