+++ packages/sanity/src/core/config/resolveDefaultPlugins.ts
@@ -1,6 +1,6 @@
 import {MEDIA_LIBRARY_NAME, mediaLibrary} from '../../media-library/plugin'
 import {CANVAS_INTEGRATION_NAME, canvasIntegration} from '../canvas/canvasIntegrationPlugin'
-import {comments} from '../comments/plugin'
+import {comments} from '../comments-legacy/plugin'
 import {releases, RELEASES_NAME} from '../releases/plugin'
 // oxlint-disable-next-line no-restricted-imports
 import {SCHEDULED_PUBLISHING_NAME, scheduledPublishing} from '../scheduled-publishing/plugin'
