+++ packages/sanity/src/core/tasks/components/activity/TasksActivityCommentInput.tsx
@@ -5,8 +5,8 @@ import {
   CommentInput,
   type CommentInputHandle,
   type CommentInputProps,
-} from '../../../comments/components/pte/comment-input/CommentInput'
-import {hasCommentMessageValue} from '../../../comments/helpers'
+} from '../../../comments-legacy/components/pte/comment-input/CommentInput'
+import {hasCommentMessageValue} from '../../../comments-legacy/helpers'
 import {useTranslation} from '../../../i18n/hooks/useTranslation'
 import {useTasksEnabled} from '../../context/enabled/useTasksEnabled'
 import {tasksLocaleNamespace} from '../../i18n'
