+++ packages/sanity/src/core/tasks/components/form/fields/descriptionInput/render/renderBlock.tsx
@@ -1,4 +1,4 @@
-import {type CommentInputRenderBlock} from '../../../../../../comments/components/pte/comment-input/CommentInput'
+import {type CommentInputRenderBlock} from '../../../../../../comments-legacy/components/pte/comment-input/CommentInput'
 import {DescriptionInputBlock} from '../blocks/DescriptionInputBlock'
 
 export const renderBlock: CommentInputRenderBlock = (props) => {
