+++ packages/sanity/src/core/tasks/components/form/tasksFormBuilder/FormEdit.tsx
@@ -13,7 +13,7 @@ import {Flex, Box} from 'ui5'
 import {MenuButton} from '../../../../../ui-components/menuButton/MenuButton'
 import {MenuItem} from '../../../../../ui-components/menuItem/MenuItem'
 import {TooltipDelayGroupProvider} from '../../../../../ui-components/tooltipDelayGroupProvider/TooltipDelayGroupProvider'
-import {CommentsProvider} from '../../../../comments/context/comments/CommentsProvider'
+import {CommentsProvider} from '../../../../comments-legacy/context/comments/CommentsProvider'
 import {ContextMenuButton} from '../../../../components/contextMenuButton/ContextMenuButton'
 import {LoadingBlock} from '../../../../components/loadingBlock/LoadingBlock'
 import {set} from '../../../../form/patch/patch'
